Revert "Merge branch 'pmphuc/feature/fsel-4885_optimize-answer-table-size' into 'hotfix-main'"
This reverts merge request !5699

+++ Services/Fsel.Course/Fsel.Course.Lms.Application/Commands/PlacementTestCmd/V1i1/SavePlacementTestAnswersCommand.cs
@@ -184,7 +184,7 @@ namespace Fsel.Course.Lms.Application.Commands.PlacementTestCmd.V1i1
             return methodResult;
         }
 
-        private static PlacementTestAnswer GetPlacementTestAnswer(PlacementTestAnswer placementTestAnswer, object? answer, short correctCount, bool isAnswered, Question questionItem)
+        private static PlacementTestAnswer GetPlacementTestAnswer(PlacementTestAnswer placementTestAnswer, object? answer, int correctCount, bool isAnswered, Question questionItem)
         {
             placementTestAnswer.Answer = answer;
             placementTestAnswer.CorrectCount = correctCount;
